--- server/src/main/java/com/zstwp/mans/security/RsaKeyProperties.java
@@ -1,11 +1,0 @@
-//package com.zstwp.mans.config;
-//
-//import org.springframework.boot.context.properties.ConfigurationProperties;
-//
-//import java.security.interfaces.RSAPrivateKey;
-//import java.security.interfaces.RSAPublicKey;
-//
-//@ConfigurationProperties(prefix = "rsa")
-//public record RsaKeyProperties(RSAPublicKey publicKey, RSAPrivateKey privateKey) {
-//
-//}
